+++ packages/StoryblokDotNet/ContentDelivery/StoryResponse.cs
@@ -4,7 +4,7 @@ public class StoryResponse<T>
     where T : StoryBlock
 {
     public required Story<T> Story { get; set; }
-    public int CV { get; set; }
+    public long CV { get; set; }
     // TODO: rels
     // TODO: links
     public Guid[]? RelUuids { get; set; }
